Version bump 1.0.2

+++ app/src/main/java/com/database/movie/data_layer/model/Movie.java
@@ -191,7 +191,7 @@ public class Movie {
         setVote_count(cursor.getInt(i));
 
         i = cursor.getColumnIndexOrThrow(IMovie.Columns.ID);
-        setId(cursor.getInt(i));
+        setId(cursor.getLong(i));
 
         i = cursor.getColumnIndexOrThrow(IMovie.Columns.VIDEO);
         setVideo(Boolean.valueOf(cursor.getString(i)));
